Move AST representation into a new file
